Add initial version of a PHP program to create snippets
We will call this program from Elisp to create snippets on the fly,
based on PHP function names that we give to it.  Currently the program
does not create a useful snippet.  All it will do is return an empty
parameter list if it receives the name of an existing function.