Add newly defined templates to YASnippet’s php-mode menu
This provides an Elisp function that creates snippets and associates
them with php-mode by registering them with YASnippet’s API.

We had to change the behavior of payas/create-template-string.  It
returns a string but later we want to use yas--define-template to
register the snippet it creates.  The problem is the latter function
expects the new snippet to be in the current buffer, not in a string.
So we change payas/create-template-string so that it puts the snippet
inside of the current buffer instead of returning any value.

With that done we’re able to add a simple function that creates a
temporary buffer—necessary now since payas/create-template-string no
longer does this—and stores the new snippet there, then immediately
reads the snippet from that temporary buffer, registers it with
YASnippet, and destroys the buffer.

With this patch we can execute code like

    (payas/define-template "implode")

to add a useful snippet for implode() to the YASnippet menu of
snippets for php-mode.