Comment that the PHP script expects a standard library function
The PHP script says it takes the name of a ‘PHP function’ but that can
be misunderstood as if we are saying the script accepts user-defined
functions as well.  This is not the case; the program only produces
output for functions that are part of PHP’s standard library.  So we
make a small addition to the source-code comments to reflect that.