Only define snippets that do not exist
This patch teaches the payas/define-template function to first ask
YASnippet if it already has a snippet for php-mode for the given
function name.  If so then we skip the rest of the process.  We can
reasonably assume that standard library functions are not going to
change in the middle of most programming sessions, so we do not need
to constantly redefine snippets every time.