Show variable sigils, array, and class type-hints in snippets
This adds a dollar sign before each parameter in snippets so that they
look more familiar to PHP programmers.  We also add the ‘array’
type-hint if the function expects the parameter value to be an array.
Or if it expects an object we try to show the class name for that
object, but this does not always seem to work consistently.