Get the symbol at point when creating a snippet
We call thing-at-point to get the word at point under the assumption
that word is a PHP function.  However, by asking for a word we will
not get back the full name for functions that use underscores.  For
example, if the point is at the end of ‘array_map’ then
(thing-at-point 'word) returns only ‘map’.  We need to ask for a
symbol to get the entire ‘array_map’ name as a string.