Show default values for optional parameters if possible
The reflection API lets us check to see if a parameter has a default
value, or if it accepts null if it’s an optional parameter.  So we
check for those two cases and add the default value or null if either
situation is true.  That said, this code seems to have no effect.
Testing it out on various standard library functions with defined
default values does not produce those default values in the snippets,
i.e. isDefaultValueAvailable() returns false even though the official
documentation gives a default value.