Report an error message if the PHP program will not run
This implements an Elisp function that will test the exit code of the
PHP program that actually creates the snippets.  We check that exit
code when trying to define a new template and report errors if
necessary.  Currently the only error we check for is ERROR_NOT_CLI,
which means PHP cannot run the program using the command-line SAPI for
some reason.