Teach payas/report-error to optionally accept a PHP function name
When payas/create-template causes an error then we pass the exit code
of the PHP program it calls to payas/report-error.  The PHP program
may return an error based on the function name payas/create-template
sends to it, e.g. if it does not recognize that name as a standard
library PHP function.  We can improve error reporting for scenarios
like that by letting payas/report-error accept an optional string,
which will be the same function name that caused any errors when we
called the payas/create-template function.