Generate snippets for class methods
This patch teaches the auxillary PHP program to accept an optional
second argument: the name of a class.  If given then the program can
treat the first argument, the function name, as a method name if it
fails to find any ‘plain’ function by that name.  This makes it
possible for calls such as:

    $ php Create-PHP-YASnippet.php "push" "SplQueue"

If we fail to find a method we exit with a new error.  This will help
us know on the Elisp side whether we failed, for example, to provide
the correct class name.