Put class names in the ‘#group’ directive of snippets for methods
YASnippet lets us organize snippets into sub-groups.  We already look
for an extension name to put functions into a group with that name.
If we are creating a snippet for a method then we can also use the
name of the class that defines that method.  This patch implements the
logic to create such sub-groups for methods.