Teach yas/create-php-snippet to prompt the user for a PHP class name
By using the prefix command (e.g C-u) the function will prompt the
user for the name of a class.  This lets the user create a snippet for
a class method.  We have to ask the user for the name of the class
because the package does not try to read it from the buffer.