Clarify the meaning of the ‘payas’ prefix used in the Elisp code
The source of the abbreviation may not be obvious to other
developers, so it deserves a small mention.  This patch also helps
make it clear that any function with the ‘payas’ prefix is internal to
the Elisp package and no developer should rely on those functions
outside of ‘php-auto-yasnippets.el’ file itself.