Use a method class name in the ‘group’ if there is no extension
The commentary in the PHP program says that it uses the class name for
a method to create the ‘group’ directive even in the absence of an
available extension name.  But the code logic does not match the
explanation in the comments.  This patch changes the code so that it
has the behavior that the commentary describes.