Create sub-groups for methods with a class belonging to a namespace
It is possible that

    $function->getDeclaringClass()->getName()

will return a class name belonging to a namespace.  If this is the
case then we can use the namespace to further organize the snippet we
create into more sub-groups.  All we need to do is replace PHP’s
namespace separator, the backslash, into YASnippet’s sub-group
separator, the period.