Store the comma character as part of each function parameter
If we do not need to use optional parameters we can press ‘C-d’ to
remove it from the function call.  But since the comma is outside of
the template it will remain after the deletion.  This leads to output
such as:

    filter_input($foo, $bar, , ,)

If we make the comma part of the template we get this:

    filter_input($foo, $bar   )

Removing the optional parameters leaves too much whitespace behind,
but this is still an improvement overall.