Remove leftover whitespace after deleting snippet parameters
Many PHP functions accept optional parameters and programmers will not
always need to use them.  In that situation the user can press ‘C-d’
to delete the parameter.  However, this leaves behind chunks of
whitespace that looks ugly.  This patch adds a hook that removes all
of that whitespace so that the snippet output looks natural.