Replace uses of ‘[]’ with ‘array()’ for PHP 5.3 compatibility
Because the PHP script uses the short-hand ‘[]’ notation for arrays it
requires PHP 5.4 or later.  Version 5.4 is only one year old, at this
time of writing, so developers may not have it installed.  It is more
likely that developers will at least have version 5.3.  The shorter
notation for arrays is not serious enough a reason to demand a
particular PHP version.

The impetus for this change comes from this patch from a fork of this
project:

https://github.com/trashofmasters/php-auto-yasnippets/commit/02ae1f8e135f0b09e182bfd073d098e8ac7a3328