Setup our whitespace-related hook whenever we load YASnippets
This patch first renames the function `payas-yas-hook-setup` to the
name `yas/initialize`.  That is because we now call the function after
loading YASnippets.  To me that makes the function part of the public
API and we have a convention of prefixing ‘yas’ to all public API
functions, to distinguish them from functions never meant for use
outside of the package.

The name ‘yas/initialize’ comes from similarly-named functions in
YASnippet related packages available from MELPA:

    http://melpa.milkbox.net/

Examples: clojure-snippets, go-snippets, and processing-snippets.

Steve Purcell was the one who suggested I look at packages like these
and so his name now appears in the README as a contributor for
providing the original impetus behind this change (and some others).

GitHub-Issue: 4