Do not make `payas/remove-extra-whitespace` a buffer-local hook
Now that the hook is a no-op in most situations (any time not invoked
by our `yas/create-php-snippet` function) we can register the hook
normally.  This addresses a problem I would run into whereby the hook
did nothing at all.  Admittedly this may not be the best change in the
long-term as we want to make sure our hook does not erroneously affect
any other packages that build on top of YASnippets.  But as a
short-term fix this patch looks ok.

GitHub-Issue: 4