Fix the code that checks for the correct number of arguments
When we run the PHP script with no argument we get a notice about
attempting to access an undefined offset in an array.  The script
checks for the minimum number of arguments but that code is incorrect
because we need to count() the number of arguments.  This patch fixes
that code so that running the PHP script with no argument no longer
raises a notice and immediately ends the script, aborting with the
ERROR_MISSING_ARGUMENT exit code.

For more information see:

    https://github.com/ejmr/php-auto-yasnippets/pull/2#issuecomment-33948245