Require auto-complete at compile time if available
This should solve the issues with 'symbol php-auto-yasnippets not
defined' that was previously thrown when using a byte-compiled version
of the file.

I tested this version of the code in my .emacs.d with and without the
auto-complete library available, and it worked for me, so unless I
screwed up my tests somehow, this should be good.