accept local definitions of php-auto-yasnippet-required-files
If a local value of php-auto-yasnippet-required-files is defined in an
.dir-locals.el file, it is ignored, because the call to the snippet
creation function happens in a temporary buffer, which doesn't inherit
the local values of the variables. We handle this by changing the global
value of php-auto-yasnippet-required-files when we are in the temp
buffer, and setting it back when we're done.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>